Two-factor confirmation using Time-based OTP
Per-route and global options to use regular confirmation only, use
2FC when the user is opted in, and require 2FC.

The TOTP secret is stored in Redis. To avoid storing the secret in
the Slack history, the secret is sent to users via e-mail. For
convenience, a QR code containing the same secret in a standard
format is generated as an attachment.

A "transitional" setting allows users to attempt enrollment
multiple times and to remove themselves from 2FC. The default
"secure" mode only allows a user to enroll once; to re-enroll,
or to remove themselves from 2FC, users have to ask a Lita
administrator (or a member of a dedicated group) to remove them.